requirements.txt
boto3==1.24.82
botocore==1.27.82
python-dotenv==0.21.0
python_dateutil==2.8.2
